[Chore] #181 - UIColor Extension으로 분리
